added missing labels to some plots
git-svn-id: https://svn.atnf.csiro.au/difx@7465 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/polconvert/trunk/src/PP/runpolconvert.py
@@ -124,7 +124,8 @@ def runPolConvert(label, band3=False, band6Lo=False, band6Hi=False,
             plotAnt=plotAnt, doTest=doTest)
     except Exception, ex:
         print 'Polconvert Exception'
-        shutil.rmtree(DiFXoutput)
+        if (os.path.exists(DiFXoutput)):
+            shutil.rmtree(DiFXoutput)
         os.rename(DiFXsave, DiFXoutput)
         raise ex
 
@@ -141,6 +142,7 @@ def runPolConvert(label, band3=False, band6Lo=False, band6Hi=False,
         for art in pcprods:
             if os.path.exists(art):
                 os.rename(art, outdir + '/' + art)
+        print savename + ' results moved to ' + outdir
 
 for job in djobs:
     # DiFX output dir and input files:
